Creating a clean, upstream/dev testing environment
About to move ll_notify from a built-in component to a custom component.
It seems I may want to be able to cleanly test it (with pytest)
and also try it in a clean environment. So this repo helps me do that,
or pick up where I left off if I ever need to rebuild as part of core.

* .envrc - use ./venv for testing / committing
* .gitmodules - include ll_notify as a submodule (not subtree)
* config / configuration - configure ll_notify & ll_notify_dashboard

This looks pretty good and clean. From here I can cleanup the
ll_notify repo to be standalone.

(Then do a submodule update to test here.)